:sparkles: #5 ルーティングページを追加しカウンターページを作成
