+++ src/App.css
@@ -1,10 +1,11 @@
+#root {
+  margin: 2.5em auto 0;
+  max-width: 930px;
+}
+
 .container {
   align-items: center;
   justify-content: center;
   min-height: 50vh;
   text-align: center;
-}
-
-.counter-button {
-  margin: 10px;
 }
