+++ src/App.tsx
@@ -1,36 +1,24 @@
-import React, { FC, useState } from 'react';
+import React, { FC } from 'react';
+import { Switch, Route, Redirect } from 'react-router-dom';
+
+import Counter from './components/Counter';
+import Home from './components/Home';
+
 import './App.css';
 
 const App: FC = () => {
-  const [count, setCount] = useState(0);
-
-  const increment = () => {
-    setCount(count + 1);
-  };
-
-  const decrement = () => {
-    setCount(count - 1);
-  };
-
   return (
     <div className="container">
-      <header>
-        <h1>カウンター</h1>
-      </header>
-      <div>
-        <div>
-          <h1>count</h1>
-          <h2>{count}</h2>
-        </div>
-        <div>
-          <button className="counter-button" type="button" onClick={decrement}>
-            -1
-          </button>
-          <button className="counter-button" type="button" onClick={increment}>
-            +1
-          </button>
-        </div>
-      </div>
+      <Switch>
+        <Route path="/counter">
+          <Counter />
+        </Route>
+        <Route path="/">
+          <Home />
+        </Route>
+
+        <Redirect to="/" />
+      </Switch>
     </div>
   );
 };
