+++ src/components/Counter.css
@@ -1,0 +1,3 @@
+.counter-button {
+  margin: 10px;
+}
