+++ src/components/Home.tsx
@@ -1,0 +1,12 @@
+import React, { FC } from 'react';
+import { Link } from 'react-router-dom';
+
+const Home: FC = () => {
+  return (
+    <div>
+      <Link to="/counter">Counter</Link>
+    </div>
+  );
+};
+
+export default Home;
